update nix config

+++ nixpkgs/update-home-manager.sh
@@ -2,5 +2,5 @@
 set -e
 
 nix-shell shell.nix --run "home-manager switch"
-ln -s $(pwd)/../dotfiles/dot_config/private_qutebrowser/autoconfig.yml $HOME/.config/qutebrowser/autoconfig.yml
-ln -s $(pwd)/../dotfiles/dot_config/private_qutebrowser/quickmarks $HOME/.config/qutebrowser/quickmarks
+[ -f $HOME/.config/qutebrowser/autoconfig.yml ] || ln -s $(pwd)/../dotfiles/dot_config/private_qutebrowser/autoconfig.yml $HOME/.config/qutebrowser/autoconfig.yml
+[ -f $HOME/.config/qutebrowser/quickmarks ] || ln -s $(pwd)/../dotfiles/dot_config/private_qutebrowser/quickmarks $HOME/.config/qutebrowser/quickmarks
